Breakwork
Major issues with how I was previously solving, which affected the right side of the charts roughly around 40 to 60. We detected that there was an error, but finally found what it was. Caused by a misunderstanding of what path length K should be, and some bad mental math let it go unchecked. Now path length k is always capped at # of forecast lead times.